CjrInfo/requirements.txt
xmltodict==0.12.0
requests==2.25.1
pytz==2021.1
django_redis==5.0.0
elasticsearch_dsl==7.4.0
elasticsearch==7.15.0
Django==3.2.5
djangorestframework==3.12.4
